Initial FLC support.
Can import one file tested (exported from
Aseprite), but attempting to import a larger
file taken from an old game resulted in all
frames looking the same (failure to correctly
parse delta frames).
Suspect that the parsing of line chunks is
wrong, because the code is reading colors
as bytes instead of words, but there can
be other problems as well.